+++ changelog.md
@@ -33,6 +33,9 @@ failed cypress tests
   it would be nice to give the user the control of choosing which sorting
   comes first.
 
+# 2022 04 01
+- add unit testing for Avatar
+- Add unit testing for ProfileForm
 # 2022 03 31
 - test last invoices in client profile.
 # 2022 03 30